Use correct value for an input's `name` attribute
The component library re-uses the field's `name` attribute to build
the IDs uses throughout the generated HTML, however this doesn't work
with ActiveModel object name's that include the model name (e.g.
`user[name]`).

Therefore, as a workaround, I'm including the name in the
`additional_attributes` hash. Ideally, the component library itself
would be updated to support ActiveModel-based forms out of the box.