Refactor tests to use component doubles
Previously, I was testing the HTML generated from the components under
test, which means our tests will break whenever there's a big change to
the Citizens Advice component library.

Now, I've refactored to use verifying doubles to ensure that the
components are called with the correct parameters only, leaving the
component library responsible for checking the components render as
expected.