Add dummy Rails applications in spec/dummy
Our unit tests only check that we are calling the design system
component with the expected parameters, but we need to be able to check
that a view containing our form controls is rendered as expected.

A dummy Rails application will allow us to check that our gem will
integrate correctly with Rails' models, views and controllers.