First pass adding 'cads_collection_radio_buttons'
I've implemented this in a similar way to Rails' built-in form helper
`collection_radio_buttons`, in that it takes a `collection` object, a
`value_method` and an optional `text_method`.

This allows it to be more flexible with the type of collection that can
be passed to the `RadioGroup` component, which usually expects the
collection to be an `Array` of hashes with `label` and `value` keys only.