Dates entered in a date input are persisted
Ensure that the three text inputs (one each for day, month and year)
that make up a date input element are able to populate those fields with
previously inputted values.

The actual conversion of these multiparameter assignments are handled by
`ActiveRecord::AttributeAssignment`, which is part of ActiveRecord and
NOT ActiveModel.

This required me to add the `active_record` gem to the project and
manually include the attribute assignement class into my model in order
for date conversion to work properly.

In a proper Rails application this should be included by default, but
any non-database models will still need to add 'include
ActiveRecord::AttributeAssignment' to make dates work correctly.