Date input specs only check HTML 'name' attribute
Previously, the specs checked that the date input controls have both the
correct `name` and `id` attribute. Ahead of switching the
manually-generated text inputs for the day, month and year for
`TextInput` component calls, I've updated the existing test to only care
about the `name` attribute, because that's the only one that affects
how the parameters are named when the form is submitted.