Add error CSS class to individual date part inputs
Append `cads-input--error" class to the day, month
and year inputs. For now, an error will highlight all of the fields.

Eventually, it should highlight a specific field, where applicable -
e.g. if the year was missing, only that field will get the error class.

In order to simplify things, I've extracted out an `error?` method which
is `true` when there's an error on the field in question.