Fix form controls that didn't work with error summary
In order for `cads_error_summary` to infer the HTML IDs of the elements
to target, they need to use a common naming scheme.

Now the first element in radio button and checkbox groups, as well as
the "day" field in date input, also use the same naming scheme.

Additionally, I've added the missing error handling from some of these
fields, so when a validation error occurs, the correct styling will be
applied to them.