Amend text input interface
The text input was setting the name using the top-level `name` property but with the ID value and then again in additional attributes to the correct
value. Avoid the double setting and fix the element spec.

Also passes all additional attributes through to the component in
keeping with the other field types.