Support multiple Rails versions
We have some projects running older Rails versions. Increase the version
range and update the CI workflow to allow testing against the same
supported versions as the design system components themselves.

This requires providing a shim for our use of field_id and field_name
which are only available in Rails 7+ as well as adjusting the dummy spec
app to only use configuration available in all our supported versions.